Refactor build.gradle and app/build.gradle
Some dependencies were defined in the root build.gradle file, while
others were defined in the app/build.gradle file, which had a mix of
dependencies imported from the other file and originally defined ones.

We want all dependencies to be declared in the root file and only
imported ones in the inner build.gradle file, so this commit refactors
the two files to be consistent with this rule.

This commit also fixes a small bug for which extJunit was declared
twice as an androidTestImplementation instead of once as that and
once as testImplementation in app/build.gradle.