Add feature to upload profile pic from gallery
As of now, the profile picture can only be uploaded by shooting a new
photo. This is currently done by clicking on the profile picture,
which makes a plus sign appear on top of it. Clicking it prompts the
user to shoot a new profile picture.

This commit alters the flow by adding two buttons, making the user
choose between uploading the picture from the camera or gallery.

Clicking on the profile picture does nothing after this commit, unless
a future change decides to change the behavior again.