Add profile picture cropping and rotation
As of now, when the user uploads or shoots a profile picture, the
results can be pretty inconsistent. First of all, the shot pictures
get compressed, while the uploaded ones don't. Secondly, there is no
option to crop the pictures, so the user does not have much control
over what part of the picture gets used for the profile picture. And
last but not least, with some camera apps the shot pictures are
rotated 90 degrees when visualized in the app.

This commit not only introduces uCrop for cropping and rotating the
image after uploading it or shooting it, but performs a compression
also on the uploaded image, and refactors the code to reuse as much
lines as possible between the two cases.