Add a view to show an insertion's details
Please merge the following commit on the backend before this one:
https://github.com/MatteoArella/passit-backend/pull/5

Currently, there is no way to view an insertion's details after
creating it or searching for it. This commit creates an Activity that
fulfills that role, querying the backend for the insertion (if not
present in the local DB) and displaying its info on a dedicated
screen.