Remove profile picture bitmap decoding
User profile picture is now cropped without decoding picture bitmap.
If the picture is taken from camera, a temp file is created for storing
the picture and the uri is then used as source for cropping.
If the picture is taken from storage its uri is directly used for
cropping.
The image compression is done directly with UCrop instead of doing it
inside the ProfileFragment and the upload is done using the cropped
image uri.